+++ app/tg/alert.py
@@ -13,9 +13,9 @@ def send_message(message: str) -> None:
         message: Message text to send
     """
     config = Configuration()
-    bot = telegram.Bot(config.getToken())
+    bot = telegram.Bot(config.get_token())
 
     loop = asyncio.get_event_loop()
     loop.create_task(
-        bot.send_message(text=message, chat_id=config.getAdminID(), parse_mode="HTML")
+        bot.send_message(text=message, chat_id=config.get_chat_id(), parse_mode="HTML")
     )